make sure we don't have duplicate code
