+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveDashboard/InteractiveDashboard.tsx
@@ -15,7 +15,6 @@ import {
 } from "embedding-sdk/components/private/PublicComponentWrapper";
 import { renderOnlyInSdkProvider } from "embedding-sdk/components/private/SdkContext";
 import { StyledPublicComponentWrapper } from "embedding-sdk/components/public/InteractiveDashboard/InteractiveDashboard.styled";
-import { useCommonDashboardParams } from "embedding-sdk/components/public/InteractiveDashboard/use-common-dashboard-params";
 import {
   type SdkDashboardDisplayProps,
   useSdkDashboardParams,
@@ -34,6 +33,7 @@ import { getEmbeddingMode } from "metabase/visualizations/click-actions/lib/mode
 import type { ClickActionModeGetter } from "metabase/visualizations/types";
 
 import type { DrillThroughQuestionProps } from "../InteractiveQuestion/InteractiveQuestion";
+import { useCommonDashboardParams } from "../SdkDashboard/use-common-dashboard-params";
 
 import { InteractiveDashboardProvider } from "./context";
