+++ enterprise/frontend/src/embedding-sdk/components/public/SdkDashboard/use-common-dashboard-params.tsx
@@ -1,54 +1,92 @@
 import { useCallback, useEffect, useState } from "react";
-import { usePrevious } from "react-use";
+import { usePrevious, useUnmount } from "react-use";
 
-import { useSdkDispatch } from "embedding-sdk/store";
-import { NAVIGATE_TO_NEW_CARD } from "metabase/dashboard/actions";
+import { useSdkDispatch, useSdkStore } from "embedding-sdk/store";
+import {
+  NAVIGATE_TO_NEW_CARD,
+  reset as dashboardReset,
+} from "metabase/dashboard/actions";
 import { getNewCardUrl } from "metabase/dashboard/actions/getNewCardUrl";
 import type { NavigateToNewCardFromDashboardOpts } from "metabase/dashboard/components/DashCard/types";
-import type { SdkDashboardId } from "metabase/embedding-sdk/types/dashboard";
 import * as Urls from "metabase/lib/urls";
+import { navigateBackToDashboard } from "metabase/query_builder/actions";
+import { getMetadata } from "metabase/selectors/metadata";
+import type Question from "metabase-lib/v1/Question";
+import type { DashboardId, QuestionDashboardCard } from "metabase-types/api";
 
 export const useCommonDashboardParams = ({
   dashboardId,
 }: {
-  dashboardId: SdkDashboardId;
+  dashboardId: DashboardId | null;
 }) => {
-  const [adhocQuestionUrl, setAdhocQuestionUrl] = useState<string | null>(null);
-
   const dispatch = useSdkDispatch();
+  const store = useSdkStore();
+
+  const [adhocQuestionUrl, setAdhocQuestionUrl] = useState<string | null>(null);
 
   const previousDashboardId = usePrevious(dashboardId);
 
+  useUnmount(() => {
+    dispatch(dashboardReset()); // reset "isNavigatingBackToDashboard" state
+  });
+
   useEffect(() => {
     if (previousDashboardId && dashboardId !== previousDashboardId) {
+      dispatch(dashboardReset()); // reset "isNavigatingBackToDashboard" state
       setAdhocQuestionUrl(null);
     }
   }, [dashboardId, dispatch, previousDashboardId]);
 
-  const onNavigateBackToDashboard = useCallback(() => {
-    setAdhocQuestionUrl(null);
-  }, []);
+  const handleNavigateToNewCardFromDashboard = useCallback(
+    ({
+      nextCard,
+      previousCard,
+      dashcard,
+      objectId,
+    }: NavigateToNewCardFromDashboardOpts) => {
+      const state = store.getState();
+      const metadata = getMetadata(state);
+      const { dashboards, parameterValues } = state.dashboard;
+      const dashboard = dashboardId && dashboards[dashboardId];
 
-  const onNavigateToNewCardFromDashboard = useCallback(
-    (opts: NavigateToNewCardFromDashboardOpts) => {
-      const url = getNewCardUrl(opts);
-      const questionUrl = Urls.question(null, { hash: url });
+      if (dashboard) {
+        const url = getNewCardUrl({
+          metadata,
+          dashboard,
+          parameterValues,
+          nextCard,
+          previousCard,
+          dashcard: dashcard as QuestionDashboardCard,
+          objectId,
+        });
 
-      dispatch({ type: NAVIGATE_TO_NEW_CARD, payload: { url } });
-      setAdhocQuestionUrl(questionUrl);
+        if (url) {
+          dispatch({ type: NAVIGATE_TO_NEW_CARD, payload: { dashboardId } });
+          setAdhocQuestionUrl(url);
+        }
+      }
     },
-    [dispatch],
+    [dashboardId, dispatch, store],
   );
 
-  const onEditQuestion = useCallback((questionId: number) => {
-    const url = Urls.question(questionId);
-    setAdhocQuestionUrl(url);
-  }, []);
+  const handleNavigateBackToDashboard = useCallback(() => {
+    dispatch(navigateBackToDashboard(dashboardId)); // set global state for cases when navigate back from question with empty results
+
+    setAdhocQuestionUrl(null);
+  }, [dashboardId, dispatch]);
+
+  const onEditQuestion = useCallback(
+    (question: Question) => {
+      dispatch({ type: NAVIGATE_TO_NEW_CARD, payload: { dashboardId } });
+      setAdhocQuestionUrl(Urls.question(question.card()));
+    },
+    [dashboardId, dispatch],
+  );
 
   return {
     adhocQuestionUrl,
-    onNavigateBackToDashboard,
-    onNavigateToNewCardFromDashboard,
+    onNavigateBackToDashboard: handleNavigateBackToDashboard,
     onEditQuestion,
+    onNavigateToNewCardFromDashboard: handleNavigateToNewCardFromDashboard,
   };
 };
