--- frontend/src/metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/ExportDashboardAsPdf.tsx
@@ -1,12 +1,0 @@
-import { useDashboardContext } from "metabase/dashboard/context";
-
-import { ExportAsPdfButton } from "../buttons/ExportAsPdfButton";
-
-export function ExportDashboardAsPdf() {
-  const { dashboard, downloadsEnabled } = useDashboardContext();
-
-  return (
-    dashboard &&
-    downloadsEnabled.pdf && <ExportAsPdfButton dashboard={dashboard} />
-  );
-}
