+++ frontend/src/metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/constants.ts
@@ -33,5 +33,4 @@ export const DASHBOARD_VIEW_ACTIONS = [
   DASHBOARD_ACTION.DASHBOARD_INFO,
   DASHBOARD_ACTION.DASHBOARD_ACTION_MENU,
   DASHBOARD_ACTION.FULLSCREEN_ANALYTICS_DASHBOARD,
-  DASHBOARD_ACTION.DOWNLOAD_DASHBOARD_PDF,
 ];
