+++ frontend/src/metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/types.ts
@@ -1,13 +1,12 @@
 import type { ComponentType } from "react";
 
+import type { DASHBOARD_ACTION } from "metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/action-buttons";
 import type {
   DashboardFullscreenControls,
   DashboardNightModeControls,
   DashboardRefreshPeriodControls,
 } from "metabase/dashboard/types";
 import type { Collection, Dashboard } from "metabase-types/api";
-
-import type { DASHBOARD_ACTION } from "./action-types";
 
 export type DashboardActionKey = keyof typeof DASHBOARD_ACTION;
